% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incrementMatrix}
\alias{incrementMatrix}
\title{Function to increment gene matrix}
\usage{
incrementMatrix(M, events, inc)
}
\arguments{
\item{M}{data frame, colnames=patients, rownames=genes}

\item{events}{data frame. Needs columns patient and gene}

\item{inc}{number by which to increment}
}
\description{
Function to increment gene matrix
}

