% Generated by roxygen2 (4.0.2): do not edit by hand
\name{memoSort}
\alias{memoSort}
\title{Function to sort gene mutation matrix
from https://gist.github.com/dakl/5974ac1d78dac88c1c78}
\usage{
memoSort(M, sortGenes = TRUE)
}
\arguments{
\item{M}{matrix, genes as rows, samples as columns}

\item{sortGenes}{boolean wheather or not to sort genes (rows)}
}
\description{
Function to sort gene mutation matrix
from https://gist.github.com/dakl/5974ac1d78dac88c1c78
}

