% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oncoprint}
\alias{oncoprint}
\title{function to create an oncoprint plot}
\usage{
oncoprint(M, keys = list(somatic = "MUT", germline = "GERMLINE", amp = "AMP",
  del = "HOMDEL", upreg = "UP", downreg = "DOWN"), sortGenes = FALSE)
}
\arguments{
\item{M}{mutation matrix}

\item{keys}{list with the following elements: somatic, germline, amp, del, upreg, downreg}
}
\description{
function to create an oncoprint plot
}

